Separando o arquivo glsl.h
