--- include/glsl.h
@@ -1,8 +1,0 @@
-#ifndef S3D_GLSL_H
-#define S3D_GLSL_H
-
-unsigned s3d_glsl_load_program(const char *vs_path, const char *fs_path);
-unsigned s3d_glsl_load_shader(int type, const char *path);
-
-
-#endif
