+++ include/glsl/load_program.h
@@ -1,0 +1,6 @@
+#ifndef S3D_GLSL_LOAD_PROGRAM_H
+#define S3D_GLSL_LOAD_PROGRAM_H
+
+unsigned s3d_glsl_load_program(const char *vs_path, const char *fs_path);
+
+#endif
