+++ include/glsl/load_shader.h
@@ -1,0 +1,6 @@
+#ifndef S3D_GLSL_LOAD_SHADER_H
+#define S3D_GLSL_LOAD_SHADER_H
+
+unsigned s3d_glsl_load_shader(int type, const char *path);
+
+#endif
